добавил автоматическое добавление записей к бд при старте бека с учетом id

+++ backend/src/main/resources/db.migration/V2__create_stories_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS stories (
-     id SERIAL PRIMARY KEY NOT NULL,
+     id BIGINT PRIMARY KEY NOT NULL,
      bank_id VARCHAR(255),
      platform VARCHAR(255),
      font_size VARCHAR(255),
